requirements.txt
aiogram==2.7
aiohttp==3.6.2
aioredis==1.3.1
python-dotenv
pymongo==3.11.1
dnspython==2.0.0
